Seq coverage is int; by default, generate reads.

+++ simtools.py
@@ -63,7 +63,7 @@ if __name__ == '__main__':
   parser_scenario.add_argument('--seed', type=int, help='Master random seed.') 
   group_paths = parser_scenario.add_argument_group('sequencing')
   group_paths.add_argument('--seq-art-path', help='Path to ART executable.')
-  group_paths.add_argument('--seq-coverage', type=float, help='Sequencing depth.')
+  group_paths.add_argument('--seq-coverage', type=int, help='Sequencing depth.')
   parser_scenario.set_defaults(func=run_gen_scenario)  
 
   if len(sys.argv) == 1:
